A complete test of the keyboard subsystem and added toggling of a led in the Leds library
The keyboard library works fine. The main function runs a test of the
keyboard and the leds. The test is as follows:
- The only enabled key is the Power. A single press toggles the Power
led. A long press of Power key starts a Led Test and enables all keys.
All other keystrokes till then are ignored
-  After the long press already mentioned, a single press of the other
keys toggles the associated led:
   o Purifying key (Humidity/Compressor) toggles Purifying led. A long
press of it triggers repetition.
   o Timer key toggles Timer led. A long press of it triggers
repetition.
   o Power key works as before, toggling the Power led. It does not
support repetition but it supports long press
   o Speed key toggles the Low Speed led. Does not support repetition or
long pressing.
   o Anion key toggles Anion led. It does not support repetition or long
pressing, either
   o Anion + Speed at the same time toggles the High Speed led.
   o Any other key combination is filtered out by the keyboard library,
but in case of allowing one of the other ckey ombinations/modes the Tank
Full led is toggled. This should never happen, though!