I2C Library seems ready
I2C contains the necessary structure to handle the I2C hardware
subsystem as a Single Master in the bus. It is heavily based on
interrupts and interaction with a main thread to be able to handle all
situations
Now there must be a library for the I2C sensor, so we can test it.