Non Scan Leds are now synchronized to Scanned ones
The scanning of the leds and displays is at a rate of 60Hz. That means
that a Led is flashing at that rate with duty cycle of 0.33, giving a
specified luminocity. On the other hand, there are two leds that do not
take place at the scanning process; the non-scanning leds. They just
light on or off without this 60 Hz rate/0.33 duty cycle flashing. Their
luminocity is very high compared to the others.
By flashing those two leds at the same rate and duty cycle as the
others, all the leds on the panel have the same luminocity.