NTC Library seems ready!
The NTC library is tested. By placing a breakpoint at the NOP mnemonic
in NTCLoop (line 228, file main.asm) we can test the converted
temperature as read from the ADC. The NTC is a NTCLE300E3302SB of Vishay
which is placed at the high leg of a voltage divider. Next step is to
add functionality to make the ADC make a group of measurements (one of
the NTC and one of the internal temperature sensor of MSP430)