Both ADC channels as a sequence test case
Using this test case we check if the sequence of channels in ADC library
works. Using this test, both NTC and Internal Temperature Sensor are
converted to digital and Celcius values. The difference is that the
specific enabling of the NTC hardware must be performed before starting
the sequence-of-channels conversion. (That is why another tag is placed
:))