More additions for helping UART Protocol construction
In Leds library there is now functionality to get or set the whole
buffer of leds and their blinking status
In Uart library the receiving interrupt has a new mode called packet
mode. This is more convenient to receive whole packets and then wake the
system up in order to consume them.
The UART Protocol is still under construction